create_polygon_df {ggpattern}	R Documentation

Create a polygon_df object from the given coordinates

Description

code using polygon_df should not assume that the first and last point within each id are the same. i.e. they may have to manulaly set a final point equal to the initial point if that is what their graphics system desires

Usage

create_polygon_df(x, y, id = 1L)

Arguments

x, y
coordinates of polygon. not necessarily closed.

id
a numeric vector used to separate locations in x,y into multiple polygons

Value

data.frame with x, y, id columns.

Examples

  df <- create_polygon_df(x = c(0, 0, 1, 1), y = c(0, 1, 1, 0))
  is_polygon_df(df)

[Package ggpattern version 1.3.1 Index]