ggpattern {ggpattern}	R Documentation

ggpattern: 'ggplot2' Pattern Geoms

Description

Provides 'ggplot2' geoms filled with various patterns. Includes a patterned version of every 'ggplot2' geom that has a region that can be filled with a pattern. Provides a suite of 'ggplot2' aesthetics and scales for controlling pattern appearances. Supports over a dozen builtin patterns (every pattern implemented by 'gridpattern') as well as allowing custom user-defined patterns.

Package options

The following ggpattern options may be set globally via base::options():

ggpattern_array_funcs

Set custom “array” pattern functions.

ggpattern_geometry_funcs

Set custom “geometry” pattern functions.

ggpattern_res

Set new default raster resolution default (pixels per inch) for the pattern_res aesthetic.

ggpattern_use_R4.1_clipping

If TRUE use the grid clipping path feature introduced in R v4.1.0 else do a rasterGrob approximation of the clipped pattern.

ggpattern_use_R4.1_features

If TRUE sets the default for all the other ggpattern_use_R4.1_* options arguments to TRUE.

ggpattern_use_R4.1_gradients

If TRUE use the grid gradient feature introduced in R v4.1.0 else do a rasterGrob approximation of the gradient pattern.

ggpattern_use_R4.1_masks

If TRUE use the grid mask feature introduced in R v4.1.0. else do a rasterGrob approximation of the masked pattern.

ggpattern_use_R4.1_patterns

If TRUE use the grid pattern feature introduced in R v4.1.0. Available for use in writing custom patterns.

Note to use the R v4.1.0 features one needs R be (at least) version 4.1 and not all graphic devices support any/all these features. See https://www.stat.auckland.ac.nz/~paul/Reports/GraphicsEngine/definitions/definitions.html for more information on the R v4.1.0 grid features.

Author(s)

Maintainer: Trevor L. Davis trevor.l.davis@gmail.com (ORCID)

Authors:

Mike FC

ggplot2 authors

See Also

Useful links:

https://github.com/trevorld/ggpattern

https://trevorldavis.com/R/ggpattern/

Report bugs at https://github.com/trevorld/ggpattern/issues

[Package ggpattern version 1.3.1 Index]