scale_pattern_linetype {ggpattern}	R Documentation

Scale for line patterns

Description

Default line types based on a set supplied by Richard Pearson, University of Manchester. Continuous values can not be mapped to line types.

Usage

scale_pattern_linetype(..., na.value = "blank")

scale_pattern_linetype_continuous(...)

scale_pattern_linetype_discrete(..., na.value = "blank")

Arguments

...
see ggplot2::scale_linetype() for more information

na.value
The linetype to use for NA values.

Value

A ggplot2::Scale object.

Examples

  if (require("ggplot2")) {
    # 'stripe' pattern example
    df <- data.frame(level = c("a", "b", "c", 'd'), outcome = c(2.3, 1.9, 3.2, 1))
    gg <- ggplot(df) +
      geom_col_pattern(
        aes(level, outcome, pattern_fill = level, pattern_linetype = level),
        pattern_density = 0.6,
        pattern_size = 1.5,
        pattern = 'stripe',
        fill    = 'white',
        colour  = 'black',
        linewidth = 1.5
      ) +
      theme_bw(18) +
      theme(legend.position = 'none') +
      scale_pattern_linetype() +
      labs(
        title    = "ggpattern::geom_col_pattern()",
        subtitle = "pattern = 'stripe'"
      )
    plot(gg)
  }

[Package ggpattern version 1.3.1 Index]