scale_pattern_shape {ggpattern}	R Documentation

Scales for shapes, aka glyphs

Description

scale_pattern_shape maps discrete variables to six easily discernible shapes. If you have more than six levels, you will get a warning message, and the seventh and subsequence levels will not appear on the plot. Use scale_pattern_shape_manual() to supply your own values. You can not map a continuous variable to shape unless scale_pattern_shape_binned() is used. Still, as shape has no inherent order, this use is not advised..

Usage

scale_pattern_shape(..., solid = TRUE)

scale_pattern_shape_discrete(..., solid = TRUE)

scale_pattern_shape_ordinal(...)

scale_pattern_shape_continuous(...)

Arguments

...
other arguments passed to discrete_scale()

solid
Should the shapes be solid, TRUE, or hollow, FALSE?

Details

Scales for area or radius

Value

A ggplot2::Scale object.

Examples

  if (require("ggplot2")) {
    # 'pch' pattern example
    gg <- ggplot(mtcars, aes(as.factor(cyl), mpg)) +
      geom_violin_pattern(aes(fill = as.factor(cyl),
                              pattern_shape = as.factor(cyl)),
        pattern = 'pch',
        pattern_density = 0.3,
        pattern_angle = 0,
        colour  = 'black'
      ) +
      theme_bw(18) +
      theme(legend.position = 'none') +
      scale_pattern_shape() +
      labs(
        title    = "ggpattern::geom_violin_pattern()",
        subtitle = "pattern = 'pch'"
      )
    plot(gg)
  }

[Package ggpattern version 1.3.1 Index]